requirements.txt
asn1crypto==0.24.0
click==7.1.2
cryptography==2.6.1
distro-info==0.21
entrypoints==0.3
Flask==1.1.2
future==0.18.2
itsdangerous==1.1.0
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.1
numpy==1.19.0
Pillow==7.2.0
pycrypto==2.6.1
PyGObject==3.30.4
python-apt==1.8.4.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.12.0
torch==1.5.1
torchvision==0.6.1
unattended-upgrades==0.1
